+++ VirtualPets/src/customFileLoader/ToyTypeLoadFormat.java
@@ -23,8 +23,8 @@ public class ToyTypeLoadFormat implements LoadFormat {
 					attributes.get("name").substring(1, attributes.get("name").length() - 1),
 					new ImageIcon(SettingsLoader.class.getResource(
 							attributes.get("icon").substring(1, attributes.get("icon").length() - 1)
-							)),
-					Double.parseDouble(attributes.get("price")), 
+							)), 
+					Integer.parseInt(attributes.get("price")), 
 					Integer.parseInt(attributes.get("happinessGain")));
 			customObjects.add(newToyType);
 		}
